fix bug cart page - #3
